+++ src/main/java/com/github/abraham054/finalreality/model/weapon/MagicWeapon/MagicWeaponType.java
@@ -1,0 +1,5 @@
+package com.github.abraham054.finalreality.model.weapon;
+
+public enum MagicWeaponType {
+    STAFF
+}