+++ src/main/java/com/github/abraham054/finalreality/model/weapon/Weapon.java
@@ -11,37 +11,32 @@ import java.util.Objects;
 public class Weapon {
 
   private final String name;
-  private final int damage;
   private final int weight;
-  private final WeaponType type;
 
   /**
-   * Creates a weapon with a name, a base damage, speed and it's type.
-   *
-   * @see WeaponType
+   * Creates a weapon
+   * @param name
+   *     the weapon's name
+   * @param weight
+   *     the weapon's weight
    */
-  public Weapon(final String name, final int damage, final int weight,
-      final WeaponType type) {
+  public Weapon(final String name, final int weight) {
     this.name = name;
-    this.damage = damage;
     this.weight = weight;
-    this.type = type;
   }
 
+  /**
+   * Returns the weapon's name
+   * */
   private String getName() {
     return name;
   }
 
-  private int getDamage() {
-    return damage;
-  }
-
+  /**
+   * Returns the weapon's weight
+   * */
   public int getWeight() {
     return weight;
-  }
-
-  private WeaponType getType() {
-    return type;
   }
 
   @Override